requirements.txt
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
shiboken6==6.7.2
PyYAML~=6.0.2
requests~=2.32.3
pydantic~=2.9.2
urllib3~=2.2.3
curlify~=2.2.1